refactor(devtools): move devtools manager inside @wasm4/devtools package
Other changes:

- fix incorrect data buffering
- remove query-param option
